+++ projects/amdsmi/include/amd_smi/amdsmi.h
@@ -3725,6 +3725,38 @@ amdsmi_status_t amdsmi_get_gpu_vendor_name(amdsmi_processor_handle processor_han
                                            size_t len);
 
 /**
+ *  @brief Get the vram vendor string of a device.
+ *
+ *  @deprecated This API is slated for removal in a future ROCm release;
+ *  ::amdsmi_get_gpu_vram_info() should be used instead
+ *
+ *  @ingroup tagIdentQuery
+ *
+ *  @platform{gpu_bm_linux}
+ *
+ *  @details This function retrieves the vram vendor name given a processor handle
+ *  @p processor_handle, a pointer to a caller provided
+ *  char buffer @p brand, and a length of this buffer @p len, this function
+ *  will write the vram vendor of the device (up to @p len characters) to the
+ *  buffer @p brand.
+ *
+ *  If the vram vendor for the device is not found as one of the values
+ *  contained within amdsmi_get_gpu_vram_vendor, then this function will return
+ *  the string 'unknown' instead of the vram vendor.
+ *
+ *  @param[in] processor_handle a processor handle
+ *
+ *  @param[in,out] brand a pointer to a caller provided char buffer to which the
+ *  vram vendor will be written
+ *
+ *  @param[in] len the length of the caller provided buffer @p brand.
+ *
+ *  @return ::amdsmi_status_t | ::AMDSMI_STATUS_SUCCESS on success, non-zero on fail
+ */
+amdsmi_status_t amdsmi_get_gpu_vram_vendor(amdsmi_processor_handle processor_handle, char* brand,
+                                           uint32_t len);
+
+/**
  *  @brief Get the subsystem device id associated with the device with
  *  provided processor handle.
  *