+++ projects/amdsmi/py-interface/__init__.py
@@ -263,6 +263,7 @@ from .amdsmi_interface import AmdSmiEventReader
 # # Device Identification information
 from .amdsmi_interface import amdsmi_get_gpu_vendor_name
 from .amdsmi_interface import amdsmi_get_gpu_id
+from .amdsmi_interface import amdsmi_get_gpu_vram_vendor
 from .amdsmi_interface import amdsmi_get_gpu_subsystem_id
 from .amdsmi_interface import amdsmi_get_gpu_subsystem_name